Resolved some bugs that stood in the way of compiling. * In DynLib.cc there was a const string that (under linux) needed to be changed later. => Not const anymore. * In SoundManager.cc the std::max algorithm hat one argument size_t and one unsigned int. Not they're both unsigned int.

+++ code/branches/presentation2/src/orxonox/sound/SoundManager.cc
@@ -607,7 +607,7 @@ namespace orxonox
                 break;
             }
         }
-        int used = std::max(this->usedSoundSources_.size(), this->minSources_);
+        int used = std::max((unsigned int)(this->usedSoundSources_.size()), this->minSources_);
         // Subtract those we added in the statement above trough std::max
         int available = (int)this->availableSoundSources_.size() - (used - (int)this->usedSoundSources_.size());
         // Delete sources again to free resources if appropriate (more than 50% more available than used)
